finished random fight sequence
