+++ package.json
@@ -13,6 +13,7 @@
     "connect-session-sequelize": "^7.1.1",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
+    "dotenv": "^8.2.0",
     "express": "~4.16.1",
     "express-session": "^1.17.1",
     "morgan": "~1.9.1",
